--- src/components/layouts/HomeMain/HomeMain.jsx
@@ -1,15 +1,0 @@
-import TopBannerHome from '../TopBannerHome/TopBannerHome';
-import BottomBannerHome from '../BottomBannerHome/BottomBannerHome';
-import SectionHomeGroup from '../SectionHomeGroup/SectionHomeGroup';
-
-const HomeMain = () => {
-  return (
-    <main>
-      <TopBannerHome />
-      <SectionHomeGroup />
-      <BottomBannerHome />
-    </main>
-  );
-};
-
-export default HomeMain;
