+++ src/components/layouts/HomePageLayout/HomeMain/index.jsx
@@ -1,0 +1,15 @@
+import TopBannerHome from '../BannerHome/TopBannerHome/index';
+import BottomBannerHome from '../BannerHome/BottomBannerHome/index';
+import SectionHomeGroup from '../SectionHomeGroup/index';
+
+const HomeMain = () => {
+  return (
+    <main>
+      <TopBannerHome />
+      <SectionHomeGroup />
+      <BottomBannerHome />
+    </main>
+  );
+};
+
+export default HomeMain;
