--- src/pages/Community/Community.jsx
@@ -1,11 +1,0 @@
-import Header from '../../components/layouts/Header/Header';
-
-const Community = () => {
-  return (
-    <>
-      <Header />
-    </>
-  );
-};
-
-export default Community;
