+++ src/pages/Community/index.jsx
@@ -1,0 +1,11 @@
+import Header from '../../components/layouts/UI/Header/index';
+
+const Community = () => {
+  return (
+    <>
+      <Header />
+    </>
+  );
+};
+
+export default Community;
