--- src/pages/Home/Home.jsx
@@ -1,15 +1,0 @@
-import Header from '../../components/layouts/Header/Header';
-import HomeMain from '../../components/layouts/HomeMain/HomeMain';
-import Footer from '../../components/layouts/Footer/Footer';
-
-const Home = () => {
-  return (
-    <>
-      <Header />
-      <HomeMain />
-      <Footer />
-    </>
-  );
-};
-
-export default Home;
