+++ src/pages/Home/index.jsx
@@ -1,0 +1,15 @@
+import Header from '../../components/layouts/UI/Header/index';
+import HomeMain from '../../components/layouts/HomePageLayout/HomeMain/index';
+import Footer from '../../components/layouts/UI/Footer/index';
+
+const Home = () => {
+  return (
+    <>
+      <Header />
+      <HomeMain />
+      <Footer />
+    </>
+  );
+};
+
+export default Home;
