--- src/pages/Items/Items.jsx
@@ -1,15 +1,0 @@
-import Header from '../../components/layouts/Header/Header';
-import BestItem from '../../components/layouts/BestItem/BestItem';
-import TotalItem from '../../components/layouts/TotalItem/TotalItem';
-
-const Items = () => {
-  return (
-    <>
-      <Header />
-      <BestItem />
-      <TotalItem />
-    </>
-  );
-};
-
-export default Items;
