+++ src/pages/Items/index.jsx
@@ -1,0 +1,15 @@
+import Header from '../../components/layouts/UI/Header/index';
+import BestItem from '../../components/layouts/ItemsPageLayout/BestItem/index';
+import TotalItem from '../../components/layouts/ItemsPageLayout/TotalItem/index';
+
+const Items = () => {
+  return (
+    <>
+      <Header />
+      <BestItem />
+      <TotalItem />
+    </>
+  );
+};
+
+export default Items;
