+++ src/pages/Privacy/index.jsx
